pyproject.toml
[project]
name = "decisioning-system"
version = "0.1.0"
description = "Modular credit decisioning platform with explainable outcomes"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Raghav", email = "you@example.com" }, { name = "Anmol", email = "you@example.com" }]

dependencies = [
  "polars>=1.0.0",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.6",
  "duckdb>=1.0.0",
  "pyarrow>=15.0.0",
  "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "types-requests"
]

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/decisioning"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.mypy]
python_version = "3.11"
mypy_path = ["src"]
strict = false

[tool.black]
line-length = 100
target-version = ["py311"]


[dependency-groups]
dev = [
    "black>=25.12.0",
]

